:truck: Moved test and removed utils folder
