+++ src/js/documentEventHelper.js
