+++ src/js/dragHelper.js
