+++ src/js/numFormatter.js
